fixed Fast PWM

+++ mainsreq/mainsreq.h
@@ -4,6 +4,6 @@
 #include <avr/io.h>
 
 void init_PWM(void);
-void PWM(uint16_t duty_cycle);
+void PWM(uint8_t duty_cycle);
 
 #endif /* pwm_header_h */
